[step3] refactoring - 코드 정리
