+++ README.md
@@ -41,6 +41,27 @@
 - [x] 보너스 규칙(숫자 5 + 보너스 1개)에 부합하는 경우 보너스 Prize을 반환한다.
 - [x] 통계데이터에 보너스 문구를 추가한다.
 
+## 역할
+- Extractor
+  - 로또의 번호 생성의 역할을 담당한다.
+- InputView 
+  - 입력값을 위한 View을 담당한다.
+- LottoMachine
+  - 구매의 기능을 담당하며 구매 즉시 생성된 로또 티켓을 반환한다.
+- LottoNumber
+  - 로또 숫자중 하나의 숫자 값을 나타낸다.
+  - 로또 숫자는 1 ~ 45 숫자로 구성된다.
+- LottoTicket
+  - 6자리의 LottoNumber로 구성된 티켓 정보.
+- LottoPrize
+  - 매칭된 번호, 보너스 정보를 토대로 상금 정보를 반환한다.
+- LottoRule
+  - 로또에 필요한 규칙 정보를 포함한다.
+- PurchaseMoney 
+  - 로또 구매를 위한 구매금액을 담당한다.
+- ResultView
+  - 출력을 위한 View을 담당한다.
+
 ## 프로그램 요구사항
 - [x] 모든 기능을 TDD로 구현해 단위 테스트가 존재해야 한다. 단, UI(System.out, System.in) 로직은 제외
 - [x] Enum 클래스를 적용해 프로그래밍을 구현한다.
