+++ src/main/kotlin/lotto/LottoPrize.kt
@@ -1,5 +1,19 @@
 package lotto
 
+import kotlin.math.floor
+
+data class LottoPrizes(val prizes: List<LottoPrize> = emptyList()) {
+    private val prizeEachCountMap = prizes.groupingBy { it.name }.eachCount()
+
+    val prizeResult = LottoPrize.values().filter { it.price > 0 }
+        .map { Pair(it, prizeEachCountMap.getOrDefault(it.name, 0)) }
+
+    fun earnings(purchaseMoney: PurchaseMoney): Double {
+        val total = prizes.sumOf { it.price }
+        return floor(total.toDouble() / purchaseMoney.money.toDouble() * 100) / 100
+    }
+}
+
 enum class LottoPrize(val matchCount: Int, val price: Int) {
     NONE(0, 0),
     FIRST(1, 0),
