+++ src/main/kotlin/lotto/LottoTicket.kt
@@ -2,18 +2,23 @@ package lotto
 
 import lotto.LottoRule.LOTTO_NUMBER_COUNT
 
-class LottoTickets(val lottery: List<LottoTicket> = emptyList())
+class LottoTickets(val lottoTickets: List<LottoTicket> = emptyList())
 open class LottoTicket(val numbers: List<LottoNumber>) {
     init {
-
         require(numbers.size == LOTTO_NUMBER_COUNT) {
             NOT_MATCH_NUMBER_COUNT
         }
+    }
+
+    fun matchCount(winningNumbers: List<LottoNumber>): Int {
+        return winningNumbers.intersect(numbers.toSet()).size
+    }
+
+    fun isMatchBonus(bonusNumber: LottoNumber): Boolean {
+        return bonusNumber in this.numbers
     }
 
     companion object {
         private const val NOT_MATCH_NUMBER_COUNT = "로또 번호는 $LOTTO_NUMBER_COUNT 개의 숫자로 구성되어야 합니다"
     }
 }
-
-fun LottoTicket.hasBonusNumber(bonusLottoNumber: LottoNumber): Boolean = bonusLottoNumber in this.numbers
