+++ src/main/kotlin/lotto/ResultView.kt
@@ -2,9 +2,9 @@ package lotto
 
 object ResultView {
     fun printTickets(lottoTickets: LottoTickets) {
-        println("${lottoTickets.lottery.size}개를 구매했습니다.")
+        println("${lottoTickets.lottoTickets.size}개를 구매했습니다.")
 
-        lottoTickets.lottery.forEach { lottoTicket ->
+        lottoTickets.lottoTickets.forEach { lottoTicket ->
             println(lottoTicket.numbers.map { it.number })
         }
     }
